// src/app/Models/PersonalProposal/cash-credit-loan.ts
export class CashCreditLoan {
    ID: number;
    CLIENT_ID: number;
    PROPOSAL_ID: number;
    CASH_CREDIT_TYPE: string ='T';
    PURPOSE_OF_LOAN: string 
    APPROXIMATE_DAILY_SALES: number;
    APPROXIMATE_MONTHLY_SALES: number;
    APPROXIMATE_MONTHLY_CREDIT_SALES: number;
    IS_CREDIT_SALE:any =0;
    AVERAGE_RECOVERY_PERIOD: number;
    APPROXIMATE_MONTHLY_PURCHASE: number;
    APPROXIMATE_MONTHLY_CREDIT_PURCHASE: number;
    AVERAGE_CREDIT_PERIOD: number;
    APPROXIMATE_MONTHLY_STOCK: number;
    IS_ANY_PURCHASE_FROM_SISTER_CONCERN_FIRM:boolean =false;
    IS_ANY_SALES_FROM_SISTER_CONCERN_FIRM:boolean =false;
    IS_STOCK_IS_INSURED:boolean =false;
    SUM_INSURED_AMOUNT: number;
    IS_DEBTOR_LIST_PREPARED_ON_TIME:boolean =false;
    IS_STOCK_MAINTAINED_AT_OTHER_PLACE:boolean =false;
    ADDRESS_OF_PLACE_AT_STOCK_IS_MAINTAINED:string
    IS_THERE_ANY_OBSOLUTE_DAMAGED_STOCK:boolean =false;
    AMOUNT_OF_OBSOLUTE_DAMAGED_PRODUCTS: number;
    ADDRESS_ID:number
}